src/requirements.txt
fastapi
uvicorn
streamlit
numpy
python-dotenv
groq
httpx